Delegate emails delivery to Sidekiq
Now use `deliver_later` for email delivery to be processed in background
jobs. Some specs verifying emails have been sent (using the
ActionMailer::Base.deliveries array) are not working anymore because
mails are now queued but jobs are not executed.

However verifying that the mailers are being called is enough in unit
tests so it's ok to remove those specs.