Fixing duplicate introduction in documentation
Ugly hack. We don't know why the introduction is duplicated when
search is on with faceting (only for entries with empty 'sections').

We could remove the display of the introduction - but then if we disable
search, no introduction. This hack remove the display in case of search
- the result is sent back/displayed by algolia anyway. Or just display
the introduction if search is disabled.